+++ src/components/elements/Button/types.d.ts
@@ -23,24 +23,23 @@ export interface ButtonProps
     | "success"
     | "subtle"
     | "link"
-    | "help";
-  /** custom component*/
-  customComponent?: string;
-  /** html button aria-controls*/
+    | "help"
+    | "clean";
+  /** WAI-ARIA support */
   ariaControls?: string;
-  /** html button aria-expanded*/
   ariaExpanded?: boolean;
-  /** html button aria-label*/
   ariaLabel?: string;
-  /** html button aria-haspopup*/
   ariaHaspopup?: boolean;
   /** button ref*/
   buttonRef?: (ref: HTMLElement) => {};
   /** button type attribute*/
-  type?: "button" | "submit" | "reset" | undefined;
+  type?: "button" | "submit" | "reset";
   /** Name property of a linked form that the button submits when clicked. */
   form?: string;
   /** selected button */
   isSelected?: boolean;
-  children?: React.ReactChild | string;
+  /** children */
+  children?: any;
+  /** icon button without children*/
+  iconOnly?: boolean;
 }